pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agentsdk"
version = "0.1.0"
description = "Unified SDK giving AI agents identity, memory, payments, and communication primitives"
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
authors = [{ name = "Phoenix Assistant" }]
keywords = ["agents", "sdk", "identity", "memory", "payments", "mcp"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
]
dependencies = [
    "aiosqlite>=0.19.0",
    "numpy>=1.24.0",
    "cryptography>=41.0.0",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "mcp>=1.0.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-asyncio>=0.21"]

[project.scripts]
agentsdk-mcp = "agentsdk.mcp_server:main"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
